backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
insightface>=0.7.3
faiss-cpu>=1.13.2
opencv-python>=4.8.1.78
numpy>=1.26.0
python-multipart>=0.0.6
google-cloud-storage>=3.9.0
Pillow>=10.0.0
pydantic>=2.4.2
python-dotenv>=1.0.0
